created utils.hpp and migrated some code

+++ contract/sample.app/src/sample.app.cpp
@@ -1,10 +1,10 @@
 #include "../include/sample.app.hpp"
+#include "../include/utils.hpp"
 
 using namespace sample;
 
 app::app(name s, name code, datastream<const char *> ds) : eosio::contract(s, code, ds) {}
 app::~app() {}
-
 
 void app::randfromstr(string seed_user1, string seed_user2) {
     print("\n----------------");
@@ -13,7 +13,7 @@ void app::randfromstr(string seed_user1, string seed_user2) {
     print("\nstr2: ", seed_user2.c_str());
     
     print("\n----------------");
-    
+
     auto hash1 = eosio::sha256(seed_user1.c_str(), seed_user1.length());
     auto hash2 = eosio::sha256(seed_user2.c_str(), seed_user2.length());
     randfromhash(hash1, hash2);
@@ -22,7 +22,7 @@ void app::randfromstr(string seed_user1, string seed_user2) {
 }
 
 void app::randfromhash(checksum256 hash_user1, checksum256 hash_user2) {
-    size_t num = 0;
+    uint32_t num = 0;
     hash_combine(num, sha256_to_hex(hash_user1));
     hash_combine(num, sha256_to_hex(hash_user2));
